New-folder/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
redis==4.0.2
python-dotenv==0.19.0
requests==2.26.0
pymongo==3.12.0
mysql-connector-python==8.0.26 